pyproject.toml
[project]
name = "aaa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.13.1",
    "pymongo>=4.6.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.37.0",
    "torch>=2.1.0",
    "requests>=2.31.0",
]
